Move the ES output adapter to upper level
